Got the tests working with travis.
Noticed some problems with expected return values for the cases in which
the given resource didn't exist. Now it returns 404 if the resource is
missing before doing the permission checks and returning a 401 or 403.

The reason behind this, is the resource needs to fetched before calling the
islandora access_hooks. Which determine the users access.

Also updated the README so that the default policies used by islandora_tomcat
are the same the user is instructed to remove and use.